Improve Customer and Validated request smart constructors
By allowing only a single means to create a 'Customer' and a 'Validated Request' - thus via their constructor functions,
we can bake in our business requirement right in the type system. This way anytime a customer or a validated request
is used throughout the program, we are sure that they were created via legitimate means that met the business requirments.